reflexion-agent/pyproject.toml
[tool.poetry]
name = "reflexion-agent"
version = "0.1.0"
description = ""
authors = ["Henrik Seidel <heseber+github@mailbox.org>"]
readme = "README.md"
packages = [{include = "reflexion_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
black = "^24.10.0"
ruff = "^0.8.6"
isort = "^5.13.2"
langchain = "^0.3.14"
langchain-openai = "^0.2.14"
langchain-anthropic = "^0.3.1"
langgraph = "^0.2.61"
langchain-community = "^0.3.14"
pydantic = "^2.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
